requirements.txt
asgiref==3.5.2; python_version >= "3.7"
attrs==21.4.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
django==3.2; python_version >= "3.6"
djangorestframework-simplejwt==5.2.0; python_version >= "3.7"
djangorestframework==3.13.1; python_version >= "3.6"
drf-spectacular==0.22.1; python_version >= "3.6"
inflection==0.5.1; python_version >= "3.6"
jsonschema==4.7.2; python_version >= "3.7"
psycopg2-binary==2.9.3; python_version >= "3.6"
pyjwt==2.4.0; python_version >= "3.7"
pyrsistent==0.18.1; python_version >= "3.7"
python-dotenv==0.20.0; python_version >= "3.5"
pytz==2022.1; python_version >= "3.7"
pyyaml==6.0; python_version >= "3.6"
sqlparse==0.4.2; python_version >= "3.7"
uritemplate==4.1.1; python_version >= "3.6"
